pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "streamlit_notebook"
version = "0.3.6"
description = "A notebook interface for Streamlit."
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Baptiste Ferrand", email = "bferrand.maths@gmail.com" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]
dependencies = [
    "streamlit-code-editor",
    "streamlit>=1.52.0",
    "pynteract>=0.1.1",
    "python-dotenv",
    "filetype",
    "pydub",
    "audioop-lts; python_version>='3.13'",
    "modict>=0.3.3"
]

[project.optional-dependencies]
# Document processing and web scraping utilities (for get_text.py)
documents = [
    "requests",
    "beautifulsoup4",
    "lxml",
    "trafilatura",
    "PyPDF2",
    "python-docx",
    "odfpy",
    "pillow",
    "selenium",
    "get-gecko-driver",
    "pathspec"
]

# AI agent core (LLM integration, embeddings, tools)
agent = [
    "openai",
    "numpy",
    "pydantic",
    "tiktoken",
    "regex",
    "pyyaml",
    "prompt-toolkit"
]

# Full agent with document processing capabilities
agent-full = [
    "streamlit_notebook[agent]",
    "streamlit_notebook[documents]"
]

# Full data science stack (plotting, ML, stats)
datascience = [
    "matplotlib",
    "numpy",
    "seaborn",
    "pandas",
    "pillow",
    "openpyxl",
    "requests",
    "graphviz",
    "altair",
    "plotly",
    "bokeh",
    "pydeck",
    "scipy",
    "sympy",
    "scikit-learn",
    "vega-datasets"
]

# Everything (agent + documents + datascience)
full = [
    "streamlit_notebook[agent-full]",
    "streamlit_notebook[datascience]"
]

[project.urls]
Homepage = "https://github.com/B4PT0R/streamlit_notebook"

[project.entry-points."console_scripts"]
st_notebook = "streamlit_notebook.launch_app:main"

[tool.setuptools.package-data]
streamlit_notebook = [
    "app_images/*",
    "demo_notebooks/*",
    "launch_app.py",
    "agent/prompts/*",
    "agent/default_gitignore"
]

[tool.setuptools.packages.find]
include = ["streamlit_notebook*"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "streamlit.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "asttokens.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "code_editor.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "matplotlib.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "seaborn.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "plotly.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "bokeh.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pydeck.*"
ignore_missing_imports = true
